Materialrework

+++ Lamp/src/Platform/Windows/WindowsWindow.cpp
@@ -190,9 +190,9 @@ namespace Lamp
 	}
 	void WindowsWindow::SetSize(const glm::vec2& size)
 	{
-		glfwSetWindowSize(m_pWindow, size.x, size.y);
+		glfwSetWindowSize(m_pWindow, (uint32_t)size.x, (uint32_t)size.y);
 
-		WindowResizeEvent resize(size.x, size.y);
+		WindowResizeEvent resize((uint32_t)size.x, (uint32_t)size.y);
 		Lamp::Application::Get().OnEvent(resize);
 	}
 }
